# cfx-api/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/trends/top": {
      "get": {
        "operationId": "TrendsController_top",
        "parameters": [
          {
            "name": "platform",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "trends"
        ]
      }
    },
...
    "/trends/pinterest": {
      "get": {
        "operationId": "TrendsController_pinterest",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "trends"
        ]
      }
    },
    "/trends/google": {
      "get": {
        "operationId": "TrendsController_google",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "trends"
        ]
      }
    },
    "/trends/pinterest/proxy": {
      "get": {
        "operationId": "TrendsController_proxyPinterest",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "trends"
        ]
      }
    },
...
  },
  "info": {
    "title": "Circular Fashion Exchange API",
    "description": "OpenAPI export",
    "version": "0.2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}